use global variable to eliminate force thread termination
