+++ src/ray/object_manager/object_manager.cc
@@ -28,11 +28,12 @@ namespace ray {
 
 ObjectStoreRunner::ObjectStoreRunner(const ObjectManagerConfig &config) {
   if (config.object_store_memory > 0) {
-    plasma_store_.reset(new plasma::PlasmaStoreRunner(
+    plasma::plasma_store_runner.reset(new plasma::PlasmaStoreRunner(
         config.store_socket_name, config.object_store_memory, config.huge_pages,
         config.plasma_directory, ""));
     // Initialize object store.
-    store_thread_ = std::thread(&plasma::PlasmaStoreRunner::Start, plasma_store_.get());
+    store_thread_ =
+        std::thread(&plasma::PlasmaStoreRunner::Start, plasma::plasma_store_runner.get());
     // Sleep for sometime until the store is working. This can suppress some
     // connection warnings.
     std::this_thread::sleep_for(std::chrono::microseconds(500));
@@ -40,10 +41,10 @@ ObjectStoreRunner::ObjectStoreRunner(const ObjectManagerConfig &config) {
 }
 
 ObjectStoreRunner::~ObjectStoreRunner() {
-  if (plasma_store_ != nullptr) {
-    plasma_store_->Stop();
+  if (plasma::plasma_store_runner != nullptr) {
+    plasma::plasma_store_runner->Stop();
     store_thread_.join();
-    plasma_store_.reset();
+    plasma::plasma_store_runner.reset();
   }
 }
